Move preprocessing functions from textcourpus module
